DSS-797: Adding enterprise poms
git-svn-id: https://svn.cesecore.eu/svn/signserver-restricted/trunk@5515 91c8134f-2a44-0410-ba45-dce2f532dbc6

+++ signserver/modules/maven-disable-nbprojects.sh
@@ -5,8 +5,11 @@ FILES=*
 for f in $FILES
 do
 	echo "Processing $f"
-	rm $f/nbproject.disabled -rf
-	mv $f/nbproject $f/nbproject.disabled
+	if [[ $f != *SignServer-Project* ]]
+	then
+		rm $f/nbproject.disabled -rf
+		mv $f/nbproject $f/nbproject.disabled
+	fi
 done
 
 echo "If NetBeans IDE was running with any of the SignServer projects open it needs to be restarted now to start use the POM.xml:s."
